requirements.txt
langchain==0.1.11
langchain-openai==0.0.5
langchain-community==0.0.27
langchain_experimental==0.0.53
python-dotenv==1.0.1
flask==3.0.2
openai==1.11.1
pypdf==4.0.1
sentence_transformers==2.5.1
faiss-cpu==1.8.0
text-generation==0.6.1 # for TGI
